long double version.  Added new functions to grab pixels and calculate simplepix values.  fixed call to polyid_and_weight.

+++ pymangle/cap.h
@@ -5,10 +5,10 @@
 #include "point.h"
 
 struct Cap {
-    double x;
-    double y;
-    double z;
-    double cm;
+    long double x;
+    long double y;
+    long double z;
+    long double cm;
 };
 
 struct CapVec {
