+++ pymangle/rand.c
@@ -16,15 +16,15 @@ seed_random(void) {
 
 void genrand_allsky(struct Point *pt)
 {
-    double theta, phi;
+    long double theta, phi;
     genrand_theta_phi_allsky(&theta, &phi);
     point_set_from_thetaphi(pt, theta, phi);
 }
-void genrand_range(double cthmin, double cthmax, 
-                   double phimin, double phimax,
+void genrand_range(long double cthmin, long double cthmax, 
+                   long double phimin, long double phimax,
                    struct Point *pt)
 {
-    double theta, phi;
+    long double theta, phi;
     genrand_theta_phi(cthmin, cthmax, 
                       phimin, phimax,
                       &theta, &phi);
@@ -32,25 +32,25 @@ void genrand_range(double cthmin, double cthmax,
     point_set_from_thetaphi(pt, theta, phi);
 }
 int 
-radec_range_to_costhetaphi(double ramin, double ramax, 
-                           double decmin, double decmax,
-                           double* cthmin, double* cthmax, 
-                           double* phimin, double* phimax)
+radec_range_to_costhetaphi(long double ramin, long double ramax, 
+                           long double decmin, long double decmax,
+                           long double* cthmin, long double* cthmax, 
+                           long double* phimin, long double* phimax)
 {
 
     if (ramin < 0.0 || ramax > 360.0) {
-        wlog("ra range must be in [0,360] got [%.16g,%.16g]",
+        wlog("ra range must be in [0,360] got [%.18Lg,%.18Lg]",
                 ramin,ramax);
         return 0;
     }
     if (decmin < -90.0 || decmax > 90.0) {
-        wlog("dec range must be in [-90,90] got [%.16g,%.16g]",
+        wlog("dec range must be in [-90,90] got [%.18Lg,%.18Lg]",
              decmin,decmax);
         return 0;
     }
 
-    *cthmin = cos((90.0-decmin)*D2R);
-    *cthmax = cos((90.0-decmax)*D2R);
+    *cthmin = cosl((90.0-decmin)*D2R);
+    *cthmax = cosl((90.0-decmax)*D2R);
     *phimin = ramin*D2R;
     *phimax = ramax*D2R;
     return 1;
@@ -62,16 +62,16 @@ radec_range_to_costhetaphi(double ramin, double ramax,
  * constant in cos(theta)
  */
 void
-genrand_theta_phi_allsky(double* theta, double* phi)
+genrand_theta_phi_allsky(long double* theta, long double* phi)
 {
-    *phi = drand48()*2*M_PI;
+    *phi = (long double) drand48()*2*M_PI;
     // this is actually cos(theta) for now
-    *theta = 2*drand48()-1;
+    *theta = 2*(long double) drand48()-1;
     
     if (*theta > 1) *theta=1;
     if (*theta < -1) *theta=-1;
 
-    *theta = acos(*theta);
+    *theta = acosl(*theta);
 }
 
 /*
@@ -81,19 +81,19 @@ genrand_theta_phi_allsky(double* theta, double* phi)
  * constant in cos(theta)
  */
 void
-genrand_theta_phi(double cthmin, double cthmax, double phimin, double phimax,
-                  double* theta, double* phi)
+genrand_theta_phi(long double cthmin, long double cthmax, long double phimin, long double phimax,
+                  long double* theta, long double* phi)
 {
 
     // at first, theta is cos(theta)
-    *phi = phimin + (phimax - phimin)*drand48();
+    *phi = phimin + (phimax - phimin)*(long double) drand48();
 
     // this is actually cos(theta) for now
-    *theta = cthmin + (cthmax-cthmin)*drand48();
+    *theta = cthmin + (cthmax-cthmin)*(long double) drand48();
     
     if (*theta > 1) *theta=1;
     if (*theta < -1) *theta=-1;
 
-    *theta = acos(*theta);
+    *theta = acosl(*theta);
 }
